IMR-150 refactor : apart genres from SelectGenrePage
